Fixes #84: SUSI does not respond until second message
