+++ src/utils/ChatMessageUtils.js
@@ -1,4 +1,3 @@
-import $ from 'jquery'
 export function convertRawMessage(rawMessage, currentThreadID) {
   return {
     ...rawMessage,
@@ -19,34 +18,21 @@ export function getCreatedMessageData(text, currentThreadID) {
   };
 };
 
-export function getSUSIMessageData(createdMessage, currentThreadID) {
+export function getSUSIMessageData(message, currentThreadID) {
   var timestamp = Date.now();
 
   let receivedMessage =  {
     id: 'm_' + timestamp,
     threadID: currentThreadID,
     authorName: 'SUSI', // hard coded for the example
+    text: message.text,
     date: new Date(timestamp),
-    text: '',
-    isRead: true
+    isRead: true,
+    responseTime: message.responseTime
   };
 
-  $.ajax({
-    url: 'http://api.asksusi.com/susi/chat.json?q='+createdMessage.text,
-    dataType: 'jsonp',
-    jsonpCallback: 'p',
-    jsonp: 'callback',
-    crossDomain: true,
-    timeout: 3000,
-    async: false,
-    success: function (response) {
-       receivedMessage.text = response.answers[0].actions[0].expression;
-      },
-    error: function(errorThrown) { 
-      console.log(errorThrown);      
-
-    } 
-  });
-  
   return receivedMessage;
+  
 }
+
+
