+++ src/utils/ChatWebAPIUtils.js
@@ -2,20 +2,14 @@
 import * as Actions from '../actions';
 import * as ChatExampleDataServer from '../ChatExampleDataServer';
 
+/*The createMessage function receives the client's message and calls for a SUSI Message creation
+using Actions.createSUSIMessage()*/
 export function createMessage(message) {
   ChatExampleDataServer.postMessage(message, createdMessage => {
     Actions.receiveCreatedMessage(createdMessage, message.id);
-  });
-  ChatExampleDataServer.postSUSIMessage(message, createdMessage => {
     Actions.createSUSIMessage(createdMessage, message.threadID);
   });
 };
-
-export function receiveSUSIMessage(message) {
-	ChatExampleDataServer.postSUSIMessage(message, receivedSUSIMessage => {
-		Actions.postSUSIMessage(message, message.id);
-	})
-}
 
 export function getAllMessages() {
   ChatExampleDataServer.getMessages(messages => {
